VoiceAssistant/requirements.txt

pyaudio
gTTS
pyautogui
pydub
speechrecognition
python-dotenv
openai
